requirements.txt
# CivicAI-Pro — pinned, tested stack
# LLM & Agents
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-openai==0.1.23
langgraph==0.2.28

# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# RAG / Vector
faiss-cpu==1.8.0
sentence-transformers==3.1.1

# HuggingFace (inference API only — no local transformers pipeline needed)
huggingface-hub==0.25.1

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
